[scanner] fix: split StatBlockFactoryModal and CustomDashboard into sub-components
Fixes #21649

StatBlockFactoryModal.tsx (851→561 lines):
- Extract types to statBlockFactoryModal.types.ts
- Extract pure functions/constants to statBlockFactoryModal.utils.ts
- Extract StatsPreview sub-component to StatBlockFactoryPreview.tsx
- Extract manage tab to StatBlockManageTab.tsx
- Remove eslint-disable max-lines comment

CustomDashboard.tsx (845→614 lines):
- Extract Card type to customDashboard/types.ts
- Extract SortableCard to customDashboard/SortableCard.tsx
- Extract DragPreviewCard to customDashboard/DragPreviewCard.tsx
- Extract empty state to customDashboard/DashboardEmptyState.tsx
- Extract delete modal to customDashboard/DashboardDeleteModal.tsx
- Remove eslint-disable max-lines comment

No behavior changes — pure structural refactor.
