+++ web/src/components/dashboard/customDashboard/types.ts
@@ -1,0 +1,7 @@
+export interface Card {
+  id: string
+  card_type: string
+  config: Record<string, unknown>
+  position: { x: number; y: number; w: number; h: number }
+  title?: string
+}
